An example subclass of IODescriptorPlayStoreBase
This reimplements the Shotgun Toolkit AppStore as a TankPlayStore

Functionality is the same, we have just removed some of the constants
and moved them to class attributes.

This is an example of implementing the only method that we need to to
get a Shotgun instance for our AppStore as well as implementing any
specific Entity type changes needed.